refactor(homebrew): card editing moves to a dedicated autosaving page
The enrichment bottom sheet becomes /homebrew/:id/cards/:card_id — a full
page with room for double-sided cards: each side's fieldset sits beside
that side's art in its printed orientation. Changes autosave on a
per-input debounce (no Save button): every change event submits the form,
a footer indicator reads saved / not-saved, invalid input shows inline
errors and simply doesn't persist until fixed, and the post-save form
rebuild renders exactly the values just typed so focused inputs are
undisturbed. The page also owns the card-shape actions (split a two-sided
card, declare a single-sided one as alt art — the declare sheet moves
here); split/declare navigate back to the project. The project page's
Edit tile button becomes a plain navigate link and its sheet plumbing is
gone.
